moved swiping to the Home page
