+++ .dart_tool/chrome-device/Default/IndexedDB/http_localhost_56324.indexeddb.leveldb/LOG.old
@@ -1,0 +1,18 @@
+2023/10/11-12:21:28.981 2f2c Creating DB C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\IndexedDB\http_localhost_56324.indexeddb.leveldb since it was missing.
+2023/10/11-12:21:28.988 2f2c Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\IndexedDB\http_localhost_56324.indexeddb.leveldb/MANIFEST-000001
+2023/10/11-12:21:29.011 3588 Level-0 table #5: started
+2023/10/11-12:21:29.017 3588 Level-0 table #5: 601 bytes OK
+2023/10/11-12:21:29.020 3588 Delete type=0 #3
+2023/10/11-12:21:29.021 2f2c Manual compaction at level-0 from '\x00\x01\x00\x00\x00' @ 72057594037927935 : 1 .. '\x00\x02\x00\x00\x00' @ 0 : 0; will stop at (end)
+2023/10/11-13:17:51.398 27d4 Level-0 table #7: started
+2023/10/11-13:17:51.402 27d4 Level-0 table #7: 1391 bytes OK
+2023/10/11-13:17:51.407 27d4 Delete type=0 #4
+2023/10/11-13:17:51.407 27d4 Manual compaction at level-0 from (begin) .. (end); will stop at (end)
+2023/10/11-13:17:51.408 27d4 Manual compaction at level-1 from (begin) .. (end); will stop at '\x00\x02\x01\x02\x013\x00[\x00D\x00E\x00F\x00A\x00U\x00L\x00T\x00]\x00!\x001\x00:\x002\x002\x006\x005\x000\x001\x008\x003\x007\x009\x002\x002\x00:\x00w\x00e\x00b\x00:\x000\x006\x003\x008\x006\x00b\x00a\x000\x00b\x00a\x001\x003\x00e\x008\x005\x00e\x00a\x000\x00f\x00d\x003\x00b' @ 68 : 1
+2023/10/11-13:17:51.408 27d4 Compacting 1@1 + 1@2 files
+2023/10/11-13:17:51.411 27d4 Generated table #8@1: 21 keys, 924 bytes
+2023/10/11-13:17:51.412 27d4 Compacted 1@1 + 1@2 files => 924 bytes
+2023/10/11-13:17:51.413 27d4 compacted to: files[ 0 0 1 0 0 0 0 ]
+2023/10/11-13:17:51.414 27d4 Delete type=2 #5
+2023/10/11-13:17:51.414 27d4 Delete type=2 #7
+2023/10/11-13:17:51.416 27d4 Manual compaction at level-1 from '\x00\x02\x01\x02\x013\x00[\x00D\x00E\x00F\x00A\x00U\x00L\x00T\x00]\x00!\x001\x00:\x002\x002\x006\x005\x000\x001\x008\x003\x007\x009\x002\x002\x00:\x00w\x00e\x00b\x00:\x000\x006\x003\x008\x006\x00b\x00a\x000\x00b\x00a\x001\x003\x00e\x008\x005\x00e\x00a\x000\x00f\x00d\x003\x00b' @ 68 : 1 .. (end); will stop at (end)
