+++ .dart_tool/chrome-device/Default/IndexedDB/http_localhost_60072.indexeddb.leveldb/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000001
