+++ .dart_tool/chrome-device/Default/IndexedDB/http_localhost_60072.indexeddb.leveldb/LOG
@@ -1,0 +1,6 @@
+2023/10/11-13:23:31.416 1ad4 Creating DB C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\IndexedDB\http_localhost_60072.indexeddb.leveldb since it was missing.
+2023/10/11-13:23:31.422 1ad4 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\IndexedDB\http_localhost_60072.indexeddb.leveldb/MANIFEST-000001
+2023/10/11-13:23:31.475 3020 Level-0 table #5: started
+2023/10/11-13:23:31.478 3020 Level-0 table #5: 690 bytes OK
+2023/10/11-13:23:31.481 3020 Delete type=0 #3
+2023/10/11-13:23:31.482 27fc Manual compaction at level-0 from '\x00\x01\x00\x00\x00' @ 72057594037927935 : 1 .. '\x00\x02\x00\x00\x00' @ 0 : 0; will stop at (end)
