+++ .dart_tool/chrome-device/Default/Local Storage/leveldb/LOG
@@ -1,3 +1,3 @@
-2023/10/05-15:39:45.817 2c84 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Local Storage\leveldb/MANIFEST-000001
-2023/10/05-15:39:45.819 2c84 Recovering log #3
-2023/10/05-15:39:45.825 2c84 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Local Storage\leveldb/000003.log 
+2023/10/11-13:23:02.978 3b68 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Local Storage\leveldb/MANIFEST-000001
+2023/10/11-13:23:02.980 3b68 Recovering log #3
+2023/10/11-13:23:02.991 3b68 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Local Storage\leveldb/000003.log
