+++ .dart_tool/chrome-device/Default/Network/Network Persistent State
@@ -1,1 +1,1 @@
-{"net":{"http_server_properties":{"servers":[{"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"server":"https://riverlegacy.org","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343554505645811","port":443,"protocol_str":"quic"}],"anonymization":["MAAAACsAAABodHRwczovL29wdGltaXphdGlvbmd1aWRlLXBhLmdvb2dsZWFwaXMuY29tAA==",false],"network_stats":{"srtt":18532},"server":"https://optimizationguide-pa.googleapis.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343562606340615","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"server":"https://content-autofill.googleapis.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343563497504036","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"server":"https://beacons5.gvt3.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343574190266377","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":45878},"server":"https://clients2.google.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343574564520658","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":35685},"server":"https://beacons.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343575686913932","port":443,"protocol_str":"quic"}],"anonymization":["GAAAABIAAABodHRwczovL2dvb2dsZS5jb20AAA==",false],"network_stats":{"srtt":42105},"server":"https://accounts.google.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343576567997526","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":39436},"server":"https://beacons3.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343575697493402","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":35926},"server":"https://www.gstatic.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343575697041822","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":36472},"server":"https://fonts.gstatic.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343577627353667","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":40504},"server":"https://firestore.googleapis.com","supports_spdy":true},{"anonymization":["FAAAAA8AAABodHRwczovL3NjZG4uY28A",false],"server":"https://i.scdn.co","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343577694208063","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":41189},"server":"https://beacons4.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343577754277585","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":62996},"server":"https://beacons2.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343577754377893","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":40023},"server":"https://beacons.gcp.gvt2.com","supports_spdy":true}],"supports_quic":{"address":"192.168.32.11","used_quic":true},"version":5},"network_qualities":{"CAASABiAgICA+P////8B":"4G","CAISABiAgICA+P////8B":"4G"}}}
+{"net":{"http_server_properties":{"servers":[{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343554505645811","port":443,"protocol_str":"quic"}],"anonymization":["MAAAACsAAABodHRwczovL29wdGltaXphdGlvbmd1aWRlLXBhLmdvb2dsZWFwaXMuY29tAA==",false],"network_stats":{"srtt":18532},"server":"https://optimizationguide-pa.googleapis.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343563497504036","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"server":"https://beacons5.gvt3.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343574190266377","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":45878},"server":"https://clients2.google.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343574564520658","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":35685},"server":"https://beacons.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343576567997526","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":39436},"server":"https://beacons3.gvt2.com","supports_spdy":true},{"anonymization":["FAAAAA8AAABodHRwczovL3NjZG4uY28A",false],"server":"https://i.scdn.co","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13343577754277585","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":62996},"server":"https://beacons2.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344085492101450","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":39151},"server":"https://content-autofill.googleapis.com","supports_spdy":true},{"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"server":"https://riverlegacy.org","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344088983081650","port":443,"protocol_str":"quic"}],"anonymization":["GAAAABIAAABodHRwczovL2dvb2dsZS5jb20AAA==",false],"network_stats":{"srtt":42359},"server":"https://accounts.google.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344088999517698","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":45786},"server":"https://fonts.gstatic.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344089000157586","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":39559},"server":"https://www.gstatic.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344089156125047","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":42379},"server":"https://beacons.gcp.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344089156214158","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":14652},"server":"https://beacons4.gvt2.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13344089156432346","port":443,"protocol_str":"quic"}],"anonymization":["FAAAABAAAABodHRwOi8vbG9jYWxob3N0",false],"network_stats":{"srtt":44520},"server":"https://firestore.googleapis.com","supports_spdy":true}],"supports_quic":{"address":"10.209.120.115","used_quic":true},"version":5},"network_qualities":{"CAASABiAgICA+P////8B":"4G","CAISABiAgICA+P////8B":"4G"}}}
