+++ .dart_tool/chrome-device/Default/Service Worker/Database/LOG
@@ -1,3 +1,3 @@
-2023/10/05-15:39:46.025 2994 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Service Worker\Database/MANIFEST-000001
-2023/10/05-15:39:46.026 2994 Recovering log #3
-2023/10/05-15:39:46.026 2994 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Service Worker\Database/000003.log 
+2023/10/11-13:23:02.807 9f8 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Service Worker\Database/MANIFEST-000001
+2023/10/11-13:23:02.808 9f8 Recovering log #3
+2023/10/11-13:23:02.808 9f8 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Service Worker\Database/000003.log
