+++ .dart_tool/chrome-device/Default/Service Worker/Database/LOG.old
@@ -1,3 +1,3 @@
-2023/10/05-14:48:06.449 3920 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Service Worker\Database/MANIFEST-000001
-2023/10/05-14:48:06.458 3920 Recovering log #3
-2023/10/05-14:48:06.460 3920 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Service Worker\Database/000003.log 
+2023/10/11-12:21:14.946 2f2c Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Service Worker\Database/MANIFEST-000001
+2023/10/11-12:21:14.948 2f2c Recovering log #3
+2023/10/11-12:21:14.949 2f2c Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Service Worker\Database/000003.log
