+++ .dart_tool/chrome-device/Default/Session Storage/LOG
@@ -1,3 +1,3 @@
-2023/10/05-15:39:46.047 2c84 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Session Storage/MANIFEST-000001
-2023/10/05-15:39:46.053 2c84 Recovering log #3
-2023/10/05-15:39:46.059 2c84 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Session Storage/000003.log 
+2023/10/11-13:23:02.974 1e54 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Session Storage/MANIFEST-000001
+2023/10/11-13:23:02.980 1e54 Recovering log #3
+2023/10/11-13:23:02.986 1e54 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Session Storage/000003.log
