+++ .dart_tool/chrome-device/Default/Session Storage/LOG.old
@@ -1,3 +1,3 @@
-2023/10/05-14:48:06.533 5040 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Session Storage/MANIFEST-000001
-2023/10/05-14:48:06.536 5040 Recovering log #3
-2023/10/05-14:48:06.544 5040 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Session Storage/000003.log 
+2023/10/11-12:21:14.949 29d4 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Session Storage/MANIFEST-000001
+2023/10/11-12:21:14.961 29d4 Recovering log #3
+2023/10/11-12:21:14.970 29d4 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Session Storage/000003.log
