+++ .dart_tool/chrome-device/Default/Site Characteristics Database/LOG
@@ -1,8 +1,8 @@
-2023/10/05-15:39:45.678 4b74 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Site Characteristics Database/MANIFEST-000001
-2023/10/05-15:39:45.679 4b74 Recovering log #10
-2023/10/05-15:39:45.679 4b74 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.35ddf0a\flutter_tools_chrome_device.8be7dae\Default\Site Characteristics Database/000010.log 
-2023/10/05-15:39:45.680 4b74 Delete type=0 #3
-2023/10/05-15:39:45.680 4b74 Delete type=2 #4
-2023/10/05-15:39:45.680 4b74 Delete type=0 #6
-2023/10/05-15:39:45.680 4b74 Delete type=2 #7
-2023/10/05-15:39:45.680 4b74 Delete type=2 #8
+2023/10/11-13:23:02.422 115c Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Site Characteristics Database/MANIFEST-000001
+2023/10/11-13:23:02.423 115c Recovering log #10
+2023/10/11-13:23:02.423 115c Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.bb941702\flutter_tools_chrome_device.828eda06\Default\Site Characteristics Database/000010.log 
+2023/10/11-13:23:02.424 115c Delete type=0 #3
+2023/10/11-13:23:02.424 115c Delete type=2 #4
+2023/10/11-13:23:02.424 115c Delete type=0 #6
+2023/10/11-13:23:02.424 115c Delete type=2 #7
+2023/10/11-13:23:02.424 115c Delete type=2 #8
