+++ .dart_tool/chrome-device/Default/Site Characteristics Database/LOG.old
@@ -1,8 +1,8 @@
-2023/10/05-14:48:06.163 5644 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Site Characteristics Database/MANIFEST-000001
-2023/10/05-14:48:06.164 5644 Recovering log #10
-2023/10/05-14:48:06.164 5644 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.5227175\flutter_tools_chrome_device.98479002\Default\Site Characteristics Database/000010.log 
-2023/10/05-14:48:06.165 5644 Delete type=0 #3
-2023/10/05-14:48:06.165 5644 Delete type=2 #4
-2023/10/05-14:48:06.165 5644 Delete type=0 #6
-2023/10/05-14:48:06.165 5644 Delete type=2 #7
-2023/10/05-14:48:06.165 5644 Delete type=2 #8
+2023/10/11-12:21:14.737 3588 Reusing MANIFEST C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Site Characteristics Database/MANIFEST-000001
+2023/10/11-12:21:14.738 3588 Recovering log #10
+2023/10/11-12:21:14.739 3588 Reusing old log C:\Users\mihai\AppData\Local\Temp\flutter_tools.8b4621e1\flutter_tools_chrome_device.2ab54360\Default\Site Characteristics Database/000010.log 
+2023/10/11-12:21:14.739 3588 Delete type=0 #3
+2023/10/11-12:21:14.739 3588 Delete type=2 #4
+2023/10/11-12:21:14.739 3588 Delete type=0 #6
+2023/10/11-12:21:14.740 3588 Delete type=2 #7
+2023/10/11-12:21:14.740 3588 Delete type=2 #8
